new enemy model
Added EnemyV2 class. This is an effort to centralise how the enemy animates. Right now, all of the lists of sprites are external to the enemy object itself, but by appending all of these as a list of sprites, the enemy class itself can potentially handle the enemy animation.  For this purpose, many animation data int lists have been added aswell as the default, full quality constructor, this is currently not yet in use. But will require further testing and improvement.