Added sound effects, tried to work up a makeshift sound engine for atleast SFX, it does not yet work
Added also advanced logging tools for audio system, it is not finding the registered sounds. There is a logical usage error between my own soundtypes, and the soundtypes it needs registered inside its own class. So the next commit needs to resolve the difference: the current ArrayID property on my sound object format, needs to be a number (no matter if it is in order or not) in order to address a specific sound effect, incase there is two sound effects sharing a name/category. The sound system proposed by AI uses its own internal record/dictrionary for them. This gap is bridgeable but it is 3am and i have work tomorrow in like, 4 hrs. oh yeah, no longer ogg-based. will attempt WAV first